format 时间.

+++ jkeep-jpa/src/main/java/com/ml/jkeep/jpa/system/vo/MenuVo.java
@@ -49,7 +49,7 @@ public class MenuVo {
     /**
      * 是否显示(1-显示,0-不显示)
      */
-    private Integer show;
+    private Byte show;
 
     private List<MenuVo> children;
 }
